DummyCandidateFinder: Rework DummyCandidateFinder
